+++ downloads.sh
@@ -78,8 +78,12 @@ cd ..
 ## VERSION: 0.4.1
 ## TYPE: file format converter
 ## SOURCE_URL: https://github.com/4dn-dcic/hic2cool
+git clone https://github.com/4dn-dcic/hic2cool.git
+cd hic2cool
+git checkout 0.4.1
+pip3 install -r requirements.txt
 pip3 install hic2cool==0.4.1
-
+cd ..
 
 ## SOFTWARE: mcool2hic
 ## COMMIT: f2d1283