+++ VERSION_HISTORY.md
@@ -3,6 +3,7 @@
 * newly introduced pairsamtools bug of not reporting lines after dedup-marked ones now fixed
 * pairsqc 0.2.3
 * `run-addfrag2pairs.sh` has an option to do nothing (-0, not adding frag info).
+* `run-juicebox-pre.sh` now has mapq filtering option and the usage has changed.
 
 ### v41
 * version updates for `bwa`, `pairix`, `cooler`, `pairsamtools`, `hic2cool`, `juicebox`